feat - add page session statistic
