+++ src/App.jsx
@@ -26,11 +26,6 @@ import RegistrationSessionHistory from "./page/HistoryRegistrationSession";
 import useRealtime from "./assets/hook/useRealTime";
 import Payment from "./page/Payment/Payment";
 import ConditionsAndServices from "./page/conditionsAndServices/conditionsAndServices";
-import AcountStatistics from "./page/AcountStatistics/AcountStatistics";
-import RequestStatistics from "./page/RequestStatistics/RequestStatistics";
-import SessionStatistics from "./page/SessionStatistics/SessionStatistics";
-import Revenues from "./page/Revenue/Revenue";
-import { motion } from "framer-motion"
 import {
   Chart as ChartJS,
   BarElement,
@@ -45,6 +40,8 @@ import {
   RadarController,
   // RadarElement,  
 } from "chart.js";
+import OverviewStatistics from "./page/OverviewStatistics/OverviewStatistics";
+import SessionStatistics from "./page/SessionStatistics/SessionStatistics";
 ChartJS.register(
   CategoryScale, // Đối với biểu đồ cột và biểu đồ đường
   LinearScale, // Đối với biểu đồ cột và biểu đồ đường
@@ -156,20 +153,12 @@ export default function App() {
           element: <ManageSession />,
         },
         {
-          path: "/dashboard/revenue",
-          element: <Revenues />,
+          path: "/dashboard/overviewStatistics",
+          element: <OverviewStatistics />,
         },
         {
-          path: "/dashboard/sessionstatistics",
+          path: "/dashboard/sessionStatistics",
           element: <SessionStatistics />,
-        },
-        {
-          path: "/dashboard/requeststatistics",
-          element: <RequestStatistics />,
-        },
-        {
-          path: "/dashboard/acountstatistics",
-          element: <AcountStatistics />,
         },
       ],
     },
