+++ src/page/SessionStatistics/SessionStatistics.jsx
@@ -1,185 +1,72 @@
 import { useEffect, useState } from "react";
-import React from 'react';
+import { Button, Modal, Form, Input, Table, Col, Row, Tooltip, Statistic } from "antd";
 import axios from "axios";
-import { Bar, Line, Pie } from "react-chartjs-2";
-import { Button, Statistic, Input, Col, Row, Select, Modal } from 'antd';
-import { UserOutlined, SettingOutlined, SmileOutlined, SyncOutlined, SearchOutlined, ConsoleSqlOutlined } from '@ant-design/icons';
-import { APIgetAllStatistics, APIgetStatisticsAcount, APIgetStatisticsRequest, APIgetStatisticsRevenue, APIgetStatisticsSession, APIgetStatisticsSessionDetail } from "../../api/api";
-const currentYear = new Date().getFullYear();
+import { Pie } from "react-chartjs-2";
+import { ContainerOutlined, DollarOutlined, HourglassOutlined, SearchOutlined, SettingOutlined, TeamOutlined } from "@ant-design/icons";
+import { APIgetAuctionSessionDetail } from "../../api/api";
+import { MarginOutlined } from "@mui/icons-material";
 
 export default function SessionStatistics() {
-  const [statisticAll, setStatisticAll] = useState();
-  const [labelChart, setLabelChart] = useState("");
-  const [isOpen, setIsOpen] = useState(false);
-  const [statisticData, setStatisticData] = useState([]);
-  const [year, setYear] = useState(currentYear);
+
   const [searchValue, setSearchValue] = useState('');
-  const [arrayDataDetail, setArrayDataDetail]  = useState([{lable :"RED",quantity:12},
-    {lable :"BLUE",quantity:12},
-    {lable :"YELLOW",quantity:12},
-    {lable :"BLACK",quantity:12},
-    {lable :"QUANG",quantity:100},
-    {lable :"PHAT",quantity:12},
-    {lable :"HOAN",quantity:12},
-    {lable :"BI",quantity:12},
-    {lable :"TRUONG",quantity:12},
-    {lable :"SANG",quantity:12},
-    {lable :"PINK",quantity:12},
-    {lable :"GREN",quantity:12}]);
-
-  const arrayData = [
-    { lable: "Session", handle: (year) => { return APIgetStatisticsSession(year) }, handlePie: () => { return APIgetStatisticsSessionDetail() } },
-    { lable: "Revenue", handle: (year) => { return APIgetStatisticsRevenue(year) } }, 
-    { lable: "Request", handle: (year) => { return APIgetStatisticsRequest(year) }, handlePie: () => { return APIgetStatisticsSessionDetail() } },   
-    { lable: "Acount", handle: (year) => { return APIgetStatisticsAcount(year) }, handlePie: () => { return APIgetStatisticsSessionDetail() } }  
-  ];
-  const years = Array.from({ length: 5 }, (v, i) => ({
-    value: currentYear - i,
-    label: `${currentYear - i}`,
-  }));
-
-  const fetchData = async () => {
-    // setIsLoading(true);
-    APIgetAllStatistics().then(rs => {
-      console.log(rs.data)
-      setStatisticAll(rs.data)
-    })
-  };
+  const [statisticAll, setStatisticAll] = useState('');
+  const [arrayDataDetail, setArrayDataDetail]  = useState([{label :"RED",quantity:12},
+    {label :"BLUE",quantity:12},
+    {label :"YELLOW",quantity:12},
+    {label :"GREN",quantity:12}]);
 
   const data = {
-    labels: ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"],
-    datasets: [
-      {
-        label: labelChart,
-        data: statisticData,
-        backgroundColor: "rgba(75, 192, 192, 0.6)",
-        borderColor: "rgba(75, 192, 192, 1)",
-        borderWidth: 0.1,
-        barThickness: 40, // Điều chỉnh chiều ngang của các cột
-      },
-      {
-        label: labelChart,
-        data: statisticData, // Dữ liệu giống như dữ liệu của cột
-        type: 'line',
-        borderColor: 'rgba(255, 99, 132, 1)', // Màu sắc của đường gấp khúc
-        borderWidth: 2,
-        fill: false,
-        tension: 0.1, // Điều chỉnh độ cong của đường
-      },
-    ],
-
-  };
-
-  const options = {
-    responsive: true,
-    plugins: {
-      legend: {
-        position: "top",
-      },
-      tooltip: {
-        callbacks: {
-          label: function (tooltipItem) {
-            return labelChart + `: ${tooltipItem.raw}`;
-          },
-        },
-      },
-    },
-    scales: {
-      x: {
-        beginAtZero: true,
-        barThickness: 20, // Chiều ngang của các cột trên trục x
-      },
-      y: {
-        beginAtZero: true,
-      },
-    },
-  };
-
-  
-    const dataPie = {
-    labels: arrayDataDetail.map((item => item.lable)),
+    labels: ["Red", "Blue", "Yellow"],
     datasets: [
       {
         label: "số lượng",
-        data:arrayDataDetail.map((item => item.quantity)),
+        data: [12, 19, 3],
         backgroundColor: [
-          '#FFD700', '#32CD32', '#FF4500', '#FF6347', '#4682B4',
-          '#8A2BE2', '#DAA520', '#DC143C', '#FF1493', '#3CB371',
-          '#00FA9A', '#B22222', '#8B0000', '#FF7F50', '#6495ED',
-          '#D2691E', '#00CED1', '#2E8B57'
-        ],
-        hoverBackgroundColor: [
-          '#FFD700', '#32CD32', '#FF4500', '#FF6347', '#4682B4',
-          '#8A2BE2', '#DAA520', '#DC143C', '#FF1493', '#3CB371',
-          '#00FA9A', '#B22222', '#8B0000', '#FF7F50', '#6495ED',
-          '#D2691E', '#00CED1', '#2E8B57'
+          "rgba(255, 99, 132, 0.2)",
+          "rgba(54, 162, 235, 0.2)",
+          "rgba(255, 206, 86, 0.2)",
         ],
         borderColor: [
-          '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF',
-          '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF',
-          '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF', '#FFFFFF',
-          '#FFFFFF', '#FFFFFF', '#FFFFFF'
+          "rgba(255, 99, 132, 1)",
+          "rgba(54, 162, 235, 1)",
+          "rgba(255, 206, 86, 1)",
         ],
         borderWidth: 1,
       },
     ],
   };
 
-  const optionsPie = {
+  const options = {
     responsive: true,
-    maintainAspectRatio: true, // Cho phép kích thước tùy chỉnh,
+    maintainAspectRatio: false, // Cho phép kích thước tùy chỉnh,
     plugins: {
       legend: {
         position: "top",
       },
     },
   };
-  const handleOnclickViewChart = async (lable) => {
-    setLabelChart(lable)
-    try {
-      const itemLable = arrayData.find(item => item.lable == lable)
-      itemLable.handle(year).then((rs) => {
-        console.log(rs.data)
-        setStatisticData(rs.data)
-      })
-    } catch (error) {
-      console.log(error)
-    }
 
+  const handleOnclickViewChart = (value) => {
+    console.log(value);
   }
-  const handleOnclickViewPie = async (lable) => {
-    setLabelChart(lable)
-    try {
-      const itemLable = arrayData.find(item => item.lable == lable)
-      itemLable.handlePie(year).then((rs) => {
-        console.log(rs.data)
-        setArrayDataDetail(rs.data)
-      })
-    } catch (error) {
-      console.log(error)
-    } finally {
-      setIsOpen(true)}
-    }
 
-  
-  useEffect(() => {
-    fetchData();
+  const handleOnclickViewPie = (value) => {
+    console.log(value);
+  }
 
-  }, []);
-  useEffect(() => {
-    // fetchData();
-    if (labelChart != null) {
-      handleOnclickViewChart(labelChart, year)
-    }
-  }, [year]);
-
-  const handleSearch = (value) => {
-    // Implement your search logic here
-    console.log("Search value:", value);
-  };
+  const handleSearchDetail = async (value) =>{
+    console.log(value);
+    APIgetAuctionSessionDetail(value).then(rs => {
+      console.log(rs.data)
+      setArrayDataDetail(rs.data)
+    }) .catch ((error) => {
+      console.log(error);
+    }) .finally(() => {
+    });
+  }
 
   const cardStyle = {
-    height: "100%",
+    height: "200px",
     width: "100%",
     padding: "20px",
     borderRadius: "10px",
@@ -189,11 +76,24 @@ export default function SessionStatistics() {
 
   const iconStyle = {
     position: "absolute",
-    top: "20px",
-    right: "20px",
-    fontSize: "24px",
+    top: "50px",
+    right: "50px",
+    fontSize: "60px",
     color: "#fff",
   };
+
+  // Tìm phần tử có label là "Total Registration"
+  const totalRegistrationItem = arrayDataDetail.find(item => item.label === "Total Registration");
+  // Lấy giá trị quantity của phần tử totalRegistrationItem
+  const totalRegistration = totalRegistrationItem ? totalRegistrationItem.quantity : 0;
+  // Tìm phần tử có label là "Total Bidding"
+  const totalBiddingItem = arrayDataDetail.find(item => item.label === "Total Bidding");
+  // Lấy giá trị quantity của phần tử totalBiddingItem
+  const totalBidding = totalBiddingItem ? totalBiddingItem.quantity : 0;
+  // Tìm phần tử có label là "Total Price"
+  const totalPriceItem = arrayDataDetail.find(item => item.label === "Total Price");
+  // Lấy giá trị quantity của phần tử totalPriceItem
+  const totalPrice = totalPriceItem ? totalPriceItem.quantity : 0;
 
   return (
     <>
@@ -211,135 +111,71 @@ export default function SessionStatistics() {
             placeholder="Search by Session ID"
             enterButton={<Button icon={<SearchOutlined />} />}
             size="middle"
-            onSearch={handleSearch}
             value={searchValue}
             onChange={(e) => setSearchValue(e.target.value)}
+            onSearch={handleSearchDetail}
           />
         </Col>
       </Row>
       <Row justify="center" gutter={20}>
-        <Col span={6}>
+        <Col span={12} >
           <div
             style={{
               ...cardStyle,
               background: "linear-gradient(to right, #B8D9D0, #3C7363)"
             }}
           >
-            <UserOutlined style={iconStyle}/>
-            <Statistic title="Total Revenue" value={statisticAll?.totalBid} valueStyle={{ fontSize: '32px' }} />
-            <Button
-              style={{ marginTop: 16 }}
-              type="primary"
-              onClick={() => {
-                // console.log("dmm")
-                handleOnclickViewChart("Revenue")
-              }}
-            >
-              View Chart
-            </Button>
+            <TeamOutlined style={iconStyle} />
+            <Statistic title="Total Auction Participants:"
+            value={totalRegistration} valueStyle={{ fontSize: '60px' }} />
+            <p style={{fontSize: "28px"}}>MEMBER</p>
           </div>
         </Col>
-        <Col span={6}>
+        <Col span={12}>
           <div
             style={{
               ...cardStyle,
               background: "linear-gradient(to right, #F2E9E9, #F2A7AD)"
             }}
           >
-            <SettingOutlined style={iconStyle} onClick={()=>{handleOnclickViewPie("Sessions")}}/>
-            <Statistic title="Total Sessions" value={statisticAll?.totalSession} valueStyle={{ fontSize: '32px' }} />
-            <Button
-              style={{ marginTop: 16 }}
-              type="primary"
-              onClick={() => {
-                // console.log("dmm")
-                handleOnclickViewChart("Session")
-              }}
-            >
-              View Chart
-            </Button>
-          </div>
-        </Col>
-        <Col span={6}>
-          <div
-            style={{
-              ...cardStyle,
-              background: "linear-gradient(to right, #97F2F3, #079DD9)"
-            }}
-          >
-            <SmileOutlined style={iconStyle} onClick={()=>{handleOnclickViewPie("Sessions")}} />
-            <Statistic title="Total Request" value={statisticAll?.totalRequest} valueStyle={{ fontSize: '32px' }} />
-            <Button
-              style={{ marginTop: 16 }}
-              type="primary"
-              onClick={() => {
-                // console.log("dmm")
-                handleOnclickViewChart("Request")
-              }}  
-            >
-              View Chart
-            </Button>
-          </div>
-        </Col>
-        <Col span={6}>
-          <div
-            style={{
-              ...cardStyle,
-              background: "linear-gradient(to right, #F3DDB3, #E08963)"
-            }}
-          >
-            <SyncOutlined style={iconStyle} onClick={()=>{handleOnclickViewPie("Sessions")}}/>
-            <Statistic title="Total Member" value={statisticAll?.totalAccount} valueStyle={{ fontSize: '32px' }} />
-            <Button
-              style={{ marginTop: 16 }}
-              type="primary"
-              onClick={() => {
-                // console.log("dmm")
-                handleOnclickViewChart("Acount")
-              }}
-            >
-              View Chart
-            </Button>
+            <ContainerOutlined style={iconStyle} />
+            <Statistic title="Total Auction Bids:" 
+            value={totalBidding} valueStyle={{ fontSize: '60px' }} />
+            <p style={{fontSize: "28px"}}>BID</p>        
           </div>
         </Col>
       </Row>
-      {labelChart &&
-        <div style={{ marginTop: "20px" }}>
-          <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between' }}>
-            <h2>{`Bar Chart with Trend Line of ${labelChart}`}</h2>
-            <Select
-              showSearch
-              style={{
-                width: 200,
-              }}
-              defaultValue={year}
-              placeholder="Select a Year"
-              optionFilterProp="label"
-              onChange={(value) => setYear(value)}
-              filterSort={(optionA, optionB) =>
-                (optionA?.label ?? '').toLowerCase().localeCompare((optionB?.label ?? '').toLowerCase())
-              }
-              options={years}
-            />
-          </div>
+      <Row justify="center" gutter={20} marginTop="20px">
+        <Col span={12}>
           <div
-          style={{
-            maxWidth: "1024px",
-            margin: "auto" 
-          }}
+            style={{
+              ...cardStyle,
+              background: "linear-gradient(to right, #97F2F3, #079DD9)",
+              marginTop: "20px",
+            }}
           >
-            <Bar data={data} options={options} />
+            <DollarOutlined style={iconStyle} />
+            <Statistic title="Total Auction Amount:" 
+            value={totalPrice} valueStyle={{ fontSize: '60px' }} /> 
+            <p style={{fontSize: "28px"}}>USD</p>      
           </div>
-        </div>
-      }
-    <Modal 
-     open={isOpen}
-     onOk={() =>setIsOpen(false)}
-     onCancel={() => setIsOpen(false)}
-    >
-      
-      <Pie data={dataPie} options={optionsPie}  />
-    </Modal>
+        </Col>
+        <Col span={12}>
+          <div
+            style={{
+              ...cardStyle,
+              background: "linear-gradient(to right, #F3DDB3, #E08963)",
+              marginTop: "20px",
+            }}
+          >
+            <HourglassOutlined style={iconStyle} />
+            <Statistic title="Total Auction Registrations:" 
+            value={totalPrice} valueStyle={{ fontSize: '60px' }} /> 
+            <p style={{fontSize: "28px"}}>MEMBER</p>
+          </div>
+        </Col>
+      </Row>
     </>
+    
   );
 }
